try a different approach to testing arrow

+++ tests/testthat/test_format_parquet.R
@@ -1,15 +1,17 @@
 context("Parquet imports/exports")
 require("datasets")
-arrow::install_arrow()
 
-test_that("Export to parquet", {
-    skip_if_not_installed("arrow")
-    expect_true(export(iris, "iris.parquet") %in% dir())
-})
+if (arrow::arrow_available()) {
+    
+    test_that("Export to parquet", {
+        skip_if_not_installed("arrow")
+        expect_true(export(iris, "iris.parquet") %in% dir())
+    })
 
-test_that("Import from parquet", {
-    skip_if_not_installed("arrow")
-    expect_true(is.data.frame(import("iris.parquet")))
-})
+    test_that("Import from parquet", {
+        skip_if_not_installed("arrow")
+        expect_true(is.data.frame(import("iris.parquet")))
+    })
 
-unlink("iris.parquet")
+    unlink("iris.parquet")
+}
